refactor: replace the three backtest exit resolvers with a policy chain
_resolve_exit, _resolve_exit_double and _resolve_structural_exit each
re-derived their own hit detection and each repeated break-even arming at
the tail, and which one ran was decided by an implicit if/elif ordering:
double won over structural, structural applied only while break-even was
unarmed. A definition setting two variant flags silently lost one of them
- structural_stop plus double_take_profit dropped the structural stop,
structural_stop plus a time cut dropped the cut.

Each rule is now one policy (Stop, StructuralStop, Target, DoubleTarget,
TimeCut, ArmBreakEven) and a definition composes an ordered chain of them
in rules.py, the only module that reads the variant flags. Ordering
carries the conventions that were previously statement order repeated
three times: the stop leads the target (FR-009), the close-measured
structural stop trails it since a target is touched intrabar first, and
break-even arming always trails everything.

Flags now compose instead of overriding each other. The four shipped
definitions build the same rules they ran before and the golden snapshot
is untouched; test_rules.py pins each shipped chain and the combinations
that used to be dropped.

The remaining test file is split by variant: test_engine (core),
test_engine_double, test_engine_structural, test_engine_time_cut,
test_parameters, test_run_range.
